Final Wilsonville Changes (#81)
* Moved swerve stuff

* Clean up

* Add controller rumbles

* Prevent thread flooding

* Add intake idle speed

* Publish boolean if drivetrain is stalling

* Update main-layout.json

* Update AdvantageScope-layout.json

* Revert "Add intake idle speed"

This reverts commit c9a01b56bb452f51bc101e2068560fa4704d36c5.

* Minor lag help

* In house testing

* I hope it doesn't lag!

* Hyped for Wilsonville!

* Remove broken vibrations

* Final (potential) lag fixes

* Add closest branch logging

* Final tweaks (again)

* Improved 3 coral auto

* Update PathPlanner PID constants
