+++ deploy/pathplanner/paths/Coral A to Boootom Tree.path
@@ -3,13 +3,13 @@
   "waypoints": [
     {
       "anchor": {
-        "x": 3.1526076158940395,
-        "y": 4.13396109271523
+        "x": 3.2108522727272732,
+        "y": 4.1795596590958946
       },
       "prevControl": null,
       "nextControl": {
-        "x": 2.9050609738893933,
-        "y": 4.099023147265855
+        "x": 2.963305630722627,
+        "y": 4.1446217136465195
       },
       "isLocked": false,
       "linkedName": "Coral A"
