+++ deploy/pathplanner/paths/Fast Left Tree to Coral L.path
@@ -1,0 +1,87 @@
+{
+  "version": "2025.0",
+  "waypoints": [
+    {
+      "anchor": {
+        "x": 1.455852272727273,
+        "y": 5.855546357615894
+      },
+      "prevControl": null,
+      "nextControl": {
+        "x": 1.9099015290951415,
+        "y": 5.855546357615894
+      },
+      "isLocked": false,
+      "linkedName": "Left Tree"
+    },
+    {
+      "anchor": {
+        "x": 3.201923076923076,
+        "y": 5.855546357615894
+      },
+      "prevControl": {
+        "x": 2.99225543493672,
+        "y": 5.991706116369651
+      },
+      "nextControl": {
+        "x": 3.4115907189094323,
+        "y": 5.719386598862137
+      },
+      "isLocked": false,
+      "linkedName": null
+    },
+    {
+      "anchor": {
+        "x": 3.717948717948717,
+        "y": 5.036410256410257
+      },
+      "prevControl": {
+        "x": 3.5303253145516433,
+        "y": 5.36138352378298
+      },
+      "nextControl": null,
+      "isLocked": false,
+      "linkedName": "Coral L"
+    }
+  ],
+  "rotationTargets": [
+    {
+      "waypointRelativePos": 1.1,
+      "rotationDegrees": -59.99999999999999
+    }
+  ],
+  "constraintZones": [],
+  "pointTowardsZones": [],
+  "eventMarkers": [
+    {
+      "name": "L4 Coral",
+      "waypointRelativePos": 0.05,
+      "endWaypointRelativePos": null,
+      "command": {
+        "type": "named",
+        "data": {
+          "name": "L4 Coral"
+        }
+      }
+    }
+  ],
+  "globalConstraints": {
+    "maxVelocity": 2.0,
+    "maxAcceleration": 3.0,
+    "maxAngularVelocity": 360.0,
+    "maxAngularAcceleration": 450.0,
+    "nominalVoltage": 12.0,
+    "unlimited": false
+  },
+  "goalEndState": {
+    "velocity": 0,
+    "rotation": -59.99999999999999
+  },
+  "reversed": false,
+  "folder": "Toddlers",
+  "idealStartingState": {
+    "velocity": 0.0,
+    "rotation": 180.0
+  },
+  "useDefaultConstraints": false
+}
