+++ deploy/pathplanner/paths/Slow Coral L to Middle Tree.path
@@ -32,12 +32,12 @@
     },
     {
       "anchor": {
-        "x": 3.1526076158940395,
-        "y": 4.13396109271523
+        "x": 3.2108522727272732,
+        "y": 4.1795596590958946
       },
       "prevControl": {
-        "x": 2.4843129139072846,
-        "y": 4.112168874172185
+        "x": 2.5425575707405184,
+        "y": 4.15776744055285
       },
       "nextControl": null,
       "isLocked": false,
