+++ robot_container.py
@@ -6,7 +6,7 @@ import commands2.button
 from commands2 import cmd, InstantCommand
 from commands2.button import CommandXboxController, Trigger
 from commands2.sysid import SysIdRoutine
-from pathplannerlib.auto import AutoBuilder, NamedCommands, PathPlannerAuto
+from pathplannerlib.auto import NamedCommands, PathPlannerAuto
 from phoenix6 import SignalLogger, swerve
 from wpilib import DriverStation, SendableChooser, XboxController, SmartDashboard, getDeployDirectory
 from wpimath.geometry import Rotation2d, Pose2d
@@ -14,12 +14,12 @@ from wpimath.units import rotationsToRadians
 
 from constants import Constants
 from generated.tuner_constants import TunerConstants
-from subsystems.auto_align import DriverAssist
 from subsystems.elevator import ElevatorSubsystem
 from subsystems.funnel import FunnelSubsystem
 from subsystems.intake import IntakeSubsystem
 from subsystems.pivot import PivotSubsystem
 from subsystems.superstructure import Superstructure
+from subsystems.swerve.requests import DriverAssist
 from subsystems.vision import VisionSubsystem
 
 
@@ -168,52 +168,26 @@ class RobotContainer:
                 lambda: self._point.with_module_direction(Rotation2d(-hid.getLeftY(), -hid.getLeftX()))
             )
         )
-        
 
         Trigger(lambda: self._driver_controller.getLeftTriggerAxis() > 0.75).whileTrue(
-            self.drivetrain.apply_request_once(
+            self.drivetrain.apply_request(
                 lambda: self._driver_assist
                 .with_velocity_x(-hid.getLeftY() * self._max_speed)
                 .with_velocity_y(-hid.getLeftX() * self._max_speed)
                 .with_rotational_rate(-self._driver_controller.getRightX() * self._max_angular_rate)
                 .with_fallback(self._field_centric)
-                .with_change_target_pose(True)
-                .with_branch_side(DriverAssist.BranchSide.LEFT)
-            )
-            .andThen(
-                self.drivetrain.apply_request(
-                lambda: self._driver_assist
-                .with_velocity_x(-hid.getLeftY() * self._max_speed)
-                .with_velocity_y(-hid.getLeftX() * self._max_speed)
-                .with_rotational_rate(-self._driver_controller.getRightX() * self._max_angular_rate)
-                .with_fallback(self._field_centric)
-                .with_change_target_pose(False)
-                .with_branch_side(DriverAssist.BranchSide.LEFT)
-            )
+                .with_target_pose(self.drivetrain.get_closest_branch(self.drivetrain.BranchSide.LEFT))
             )
         )
 
         Trigger(lambda: self._driver_controller.getRightTriggerAxis() > 0.75).whileTrue(
-            # eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee
-            self.drivetrain.apply_request_once(
+            self.drivetrain.apply_request(
                 lambda: self._driver_assist
                 .with_velocity_x(-hid.getLeftY() * self._max_speed)
                 .with_velocity_y(-hid.getLeftX() * self._max_speed)
                 .with_rotational_rate(-self._driver_controller.getRightX() * self._max_angular_rate)
                 .with_fallback(self._field_centric)
-                .with_change_target_pose(True)
-                .with_branch_side(DriverAssist.BranchSide.RIGHT)
-            )
-            .andThen(
-                self.drivetrain.apply_request(
-                lambda: self._driver_assist
-                .with_velocity_x(-hid.getLeftY() * self._max_speed)
-                .with_velocity_y(-hid.getLeftX() * self._max_speed)
-                .with_rotational_rate(-self._driver_controller.getRightX() * self._max_angular_rate)
-                .with_fallback(self._field_centric)
-                .with_change_target_pose(False)
-                .with_branch_side(DriverAssist.BranchSide.LEFT)
-            )
+                .with_target_pose(self.drivetrain.get_closest_branch(self.drivetrain.BranchSide.RIGHT))
             )
         )
 
@@ -251,7 +225,7 @@ class RobotContainer:
                 (button.whileTrue(
                     self.superstructure.set_goal_command(goal)
                     .alongWith(self.intake.set_desired_state_command(self.intake.SubsystemState.ALGAE_INTAKE)))
-                 .onFalse(self.intake.set_desired_state_command(self.intake.SubsystemState.HOLD)))
+                    .onFalse(self.intake.set_desired_state_command(self.intake.SubsystemState.ALGAE_HOLD)))
             else:
                 button.onTrue(self.superstructure.set_goal_command(goal))
