+++ subsystems/swerve/__init__.py
@@ -1,26 +1,24 @@
-from enum import Enum, auto
 import math
-from typing import Callable, overload, Self
+from enum import Enum, auto
+from typing import Callable, overload
 
-from constants import Constants
 from commands2 import Command, Subsystem
 from commands2.sysid import SysIdRoutine
 from ntcore import NetworkTableInstance
 from pathplannerlib.auto import AutoBuilder, RobotConfig
 from pathplannerlib.controller import PIDConstants, PPHolonomicDriveController
 from pathplannerlib.logging import PathPlannerLogging
-from phoenix6 import swerve, units, utils, SignalLogger, StatusCode
-from phoenix6.swerve import SwerveModule, SwerveControlParameters
-from phoenix6.swerve.requests import ApplyRobotSpeeds, FieldCentric, FieldCentricFacingAngle, ForwardPerspectiveValue, SwerveRequest
+from phoenix6 import swerve, units, utils, SignalLogger
+from phoenix6.swerve.requests import ApplyRobotSpeeds
 from phoenix6.swerve.swerve_drivetrain import DriveMotorT, SteerMotorT, EncoderT
-from phoenix6.swerve.utility.phoenix_pid_controller import PhoenixPIDController
-from phoenix6.units import meters_per_second, meter, radians_per_second
 from wpilib import DriverStation, Notifier, RobotController, Field2d, SmartDashboard
 from wpilib.sysid import SysIdRoutineLog
-from wpimath.geometry import Rotation2d, Pose2d, Translation2d
+from wpimath.geometry import Rotation2d, Pose2d
 from wpimath.kinematics import ChassisSpeeds, SwerveModuleState
 from wpimath.units import degreesToRadians
-from wpiutil import Sendable, SendableBuilder
+
+from constants import Constants
+from subsystems.swerve.requests import DriverAssist
 
 
 class SwerveSubsystem(Subsystem, swerve.SwerveDrivetrain):
@@ -35,6 +33,58 @@ class SwerveSubsystem(Subsystem, swerve.SwerveDrivetrain):
     """Blue alliance sees forward as 0 degrees (toward red alliance wall)"""
     _RED_ALLIANCE_PERSPECTIVE_ROTATION = Rotation2d.fromDegrees(180)
     """Red alliance sees forward as 180 degrees (toward blue alliance wall)"""
+
+    class BranchSide(Enum):
+        """
+        Determines which side of the reef we score on.
+        """
+        LEFT = auto()
+        RIGHT = auto()
+
+    _blue_branch_left_targets = [
+        Pose2d(3.091, 4.181, degreesToRadians(0)),  # A
+        Pose2d(3.656, 2.916, degreesToRadians(60)),  # C
+        Pose2d(5.023, 2.772, degreesToRadians(120)),  # E
+        Pose2d(5.850, 3.851, degreesToRadians(180)),  # G
+        Pose2d(5.347, 5.134, degreesToRadians(240)),  # I
+        Pose2d(3.932, 5.302, degreesToRadians(300)),  # K
+    ]
+
+    _blue_branch_right_targets = [
+        Pose2d(3.091, 3.863, degreesToRadians(0)),  # B
+        Pose2d(3.956, 2.748, degreesToRadians(60)),  # D
+        Pose2d(5.323, 2.928, degreesToRadians(120)),  # F
+        Pose2d(5.862, 4.187, degreesToRadians(180)),  # H
+        Pose2d(5.047, 5.290, degreesToRadians(240)),  # J
+        Pose2d(3.668, 5.110, degreesToRadians(300)),  # L
+    ]
+
+    _red_branch_left_targets = [
+        Pose2d(
+            Constants.FIELD_LAYOUT.getFieldLength() - pose.X(),
+            Constants.FIELD_LAYOUT.getFieldWidth() - pose.Y(),
+            pose.rotation() + Rotation2d.fromDegrees(180)
+        ) for pose in _blue_branch_left_targets
+    ]
+
+    _red_branch_right_targets = [
+        Pose2d(
+            Constants.FIELD_LAYOUT.getFieldLength() - pose.X(),
+            Constants.FIELD_LAYOUT.getFieldWidth() - pose.Y(),
+            pose.rotation() + Rotation2d.fromDegrees(180)
+        ) for pose in _blue_branch_right_targets
+    ]
+
+    _branch_targets = {
+        DriverStation.Alliance.kBlue: {
+            BranchSide.LEFT: _blue_branch_left_targets,
+            BranchSide.RIGHT: _blue_branch_right_targets,
+        },
+        DriverStation.Alliance.kRed: {
+            BranchSide.LEFT: _red_branch_left_targets,
+            BranchSide.RIGHT: _red_branch_right_targets,
+        }
+    }
 
     @overload
     def __init__(
@@ -173,33 +223,23 @@ class SwerveSubsystem(Subsystem, swerve.SwerveDrivetrain):
         # Keep track if we've ever applied the operator perspective before or not
         self._has_applied_operator_perspective = False
 
-        self._table = NetworkTableInstance.getDefault().getTable("Telemetry")
+        self._closest_left_branch, self._closest_right_branch = Pose2d(), Pose2d()
 
-        # class SwerveModuleSendable(Sendable):
-        #     def __init__(self, modules: list[SwerveModule[DriveMotorT, SteerMotorT, EncoderT]], rotation_getter: Callable[[], float]):
-        #         super().__init__()
-        #         self._modules = modules
-        #         self._get_rotation = rotation_getter
-        #
-        #     def initSendable(self, builder: SendableBuilder):
-        #         builder.setSmartDashboardType("SwerveDrive")
-        #
-        #         for i, name in enumerate(["Front Left", "Front Right", "Back Left", "Back Right"]):
-        #             builder.addDoubleProperty(f"{name} Angle", lambda: self._modules[i].get_current_state().angle.radians(), lambda _: None)
-        #             builder.addDoubleProperty(f"{name} Velocity", lambda: self._modules[i].get_current_state().speed, lambda _: None)
-        #         builder.addDoubleProperty("Robot Angle", self._get_rotation, lambda _: None)
-        # SmartDashboard.putData("Swerve Modules", SwerveModuleSendable(self.modules, lambda: (self.get_state_copy().pose.rotation() + self.get_operator_forward_direction()).radians()))
+        table = NetworkTableInstance.getDefault().getTable("Telemetry")
 
-        self._pose_pub = self._table.getStructTopic("current_pose", Pose2d).publish()
-        self._speeds_pub = self._table.getStructTopic("chassis_speeds", ChassisSpeeds).publish()
-        self._odom_freq = self._table.getDoubleTopic("odometry_frequency").publish()
-        self._module_states_pub = self._table.getStructArrayTopic("module_states", SwerveModuleState).publish()
-        self._module_targets_pub = self._table.getStructArrayTopic("module_targets", SwerveModuleState).publish()
+        self._pose_pub = table.getStructTopic("current_pose", Pose2d).publish()
+        self._speeds_pub = table.getStructTopic("chassis_speeds", ChassisSpeeds).publish()
+        self._odom_freq = table.getDoubleTopic("odometry_frequency").publish()
+        self._module_states_pub = table.getStructArrayTopic("module_states", SwerveModuleState).publish()
+        self._module_targets_pub = table.getStructArrayTopic("module_targets", SwerveModuleState).publish()
+        self._wheels_stalled_pub = table.getBooleanTopic("Wheels Stalled").publish()
 
-        self._auto_target_pub = self._table.getStructTopic("auto_target", Pose2d).publish()
-        self._auto_path_pub = self._table.getStructArrayTopic("auto_path", Pose2d).publish()
+        self._auto_target_pub = table.getStructTopic("auto_target", Pose2d).publish()
+        self._auto_path_pub = table.getStructArrayTopic("auto_path", Pose2d).publish()
         PathPlannerLogging.setLogTargetPoseCallback(lambda pose: self._auto_target_pub.set(pose))
         PathPlannerLogging.setLogActivePathCallback(lambda poses: self._auto_path_pub.set(poses))
+
+        self._closest_branch_pub = table.getStructTopic("Closest Branch", Pose2d).publish()
 
         # Swerve requests to apply during SysId characterization
         self._translation_characterization = swerve.requests.SysIdSwerveTranslation()
@@ -304,8 +344,9 @@ class SwerveSubsystem(Subsystem, swerve.SwerveDrivetrain):
                 .with_wheel_force_feedforwards_y(feedforwards.robotRelativeForcesYNewtons)
             ),
             PPHolonomicDriveController(
-                PIDConstants(7.0, 0.0, 0.0),
-                PIDConstants(7.0, 0.0, 0.0)
+                PIDConstants(10.0, 0.0, 0.0),
+                PIDConstants(10.0, 0.0, 0.0),
+                period=0.004
             ),
             config,
             lambda: DriverStation.getAlliance() == DriverStation.Alliance.kRed, # If getAlliance() is None (maybe the robot doesn't know its alliance yet), it defaults to blue. This returns True if the alliance is red, and False otherwise
@@ -323,25 +364,20 @@ class SwerveSubsystem(Subsystem, swerve.SwerveDrivetrain):
         :rtype: Command
         """
         return self.run(lambda: self.set_control(request()))
-    
-    def apply_request_once(
-            self, request: Callable[[], swerve.requests.SwerveRequest]
-    ) -> Command:
-        """
-        Returns a command that applies the specified control request to this swerve drivetrain once.
-        :param request: Lambda returning the request to apply
-        :type request: Callable[[], swerve.requests.SwerveRequest]
-        :returns: Command to run
-        :rtype: Command
-        """
-
-        return self.runOnce(lambda: self.set_control(request()))
 
     def sys_id_quasistatic(self, direction: SysIdRoutine.Direction) -> Command:
         return self._sys_id_routine_to_apply.quasistatic(direction)
 
     def sys_id_dynamic(self, direction: SysIdRoutine.Direction) -> Command:
         return self._sys_id_routine_to_apply.dynamic(direction)
+
+    def get_closest_branch(self, branch_side: BranchSide) -> Pose2d:
+        if branch_side == self.BranchSide.LEFT:
+            closest_branch =  self._closest_left_branch
+        else:
+            closest_branch = self._closest_right_branch
+        self._closest_branch_pub.set(closest_branch)
+        return closest_branch
 
     def periodic(self) -> None:
         """
@@ -353,8 +389,8 @@ class SwerveSubsystem(Subsystem, swerve.SwerveDrivetrain):
         # This allows us to correct the perspective in case the robot code restarts mid-match.
         # Otherwise, only check and apply the operator perspective if the DS is disabled.
         # This ensures driving behavior doesn't change until an explicit disable event occurs during testing.
+        alliance_color = DriverStation.getAlliance()
         if not self._has_applied_operator_perspective or DriverStation.isDisabled():
-            alliance_color = DriverStation.getAlliance()
             if alliance_color is not None:
                 self.set_operator_perspective_forward(
                     self._RED_ALLIANCE_PERSPECTIVE_ROTATION
@@ -370,6 +406,52 @@ class SwerveSubsystem(Subsystem, swerve.SwerveDrivetrain):
         self._module_states_pub.set(state.module_states)
         self._module_targets_pub.set(state.module_targets)
         self._speeds_pub.set(state.speeds)
+
+        modules_stalled = 0
+        for module in self.modules:
+            target_speed = abs(module.get_target_state().speed)
+            current_speed = abs(module.get_current_state().speed)
+
+            if 0 != target_speed > 0 and current_speed < 1e-5:
+                modules_stalled += 1
+        self._wheels_stalled_pub.set(modules_stalled >= 3)
+
+        # Calculate the closest branch
+        if alliance_color:
+            self._closest_left_branch = min(self._branch_targets[alliance_color][self.BranchSide.LEFT], key=lambda pose: self.get_distance_to_line(state.pose, pose))
+            self._closest_right_branch = min(self._branch_targets[alliance_color][self.BranchSide.RIGHT], key=lambda pose: self.get_distance_to_line(state.pose, pose))
+
+    @staticmethod
+    def get_distance_to_line(robot_pose: Pose2d, target_pose: Pose2d) -> float:
+        """
+        Find the distance from the robot to the line emanating from the target pose.
+        """
+
+        # To accomplish this, we need to find the intersection point of the line
+        # emanating from the target pose and the line perpendicular to it that passes through the robot pose.
+        # If theta is the target rotation, tan(theta) is the slope of the line from the pose. We can just call that S for the sake of solving this.
+        # Where S is the slope, (t_x, t_y) is the target position, and (r_x, r_y) is the robot position:
+        # S(x - t_x) + t_y = -1/S(x - r_x) + r_y
+        # Sx - St_x + t_y = -x/S + r_x/S + r_y
+        # Sx - St_x + t_y - r_y = (r_x - x)/S
+        # S^2x - S^2t_x + St_y - Sr_y = r_x - x
+        # S^2x + x = r_x + S^2t_x - St_y + Sr_y
+        # x(S^2 + 1) = r_x + S^2t_x - St_y + Sr_y
+        # x = (r_x + S^2t_x - St_y + Sr_y)/(S^2 + 1)
+        # We can then plug in x into our first equation to find y. This will give us the intersection point, which is the pose we want to find distance to.
+
+        slope = math.tan(target_pose.rotation().radians())
+
+        x = (robot_pose.X() + slope ** 2 * target_pose.X() - slope * target_pose.Y() + slope * robot_pose.Y()) / (slope ** 2 + 1)
+        y = slope * (x - target_pose.X()) + target_pose.Y()
+
+        possible_pose = Pose2d(x, y, target_pose.rotation())
+
+        reef_x = 4.4735 if DriverStation.getAlliance() == DriverStation.Alliance.kBlue else Constants.FIELD_LAYOUT.getFieldLength() - 4.4735
+
+        if (target_pose.X() - reef_x <= 0) == (x - reef_x <= 0):
+            return math.sqrt((possible_pose.X() - robot_pose.X()) ** 2 + (possible_pose.Y() - robot_pose.Y()) ** 2)
+        return math.inf
 
     def _start_sim_thread(self) -> None:
         """
